refactor(RenderContainer): split components on separate files
