+++ packages/retail-ui/components/RenderContainer/RenderContainer.tsx
@@ -1,38 +1,20 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
-import { CustomPortal, NativePortal } from './Portal';
-import { RootContainer } from './RootContainer';
 import { Nullable } from '../../typings/utility-types';
+import { RenderContainer as RenderContainerFallback } from './RenderContainerFallback';
+import { RenderContainer as RenderContainerNative } from './RenderContainerNative';
+import { RenderContainerProps } from './RenderContainerTypes';
 
-interface RenderContainerProps {
-  anchor?: React.ReactNode;
-  children?: React.ReactNode;
-}
+let rootId = 0;
+const HAS_BUILTIN_PORTAL = !!ReactDOM.createPortal;
 
-export class RenderContainerBase extends React.Component<RenderContainerProps> {
-  public static ids: boolean[] = [true];
+export class RenderContainer extends React.Component<RenderContainerProps> {
+  private domContainer: Nullable<HTMLElement> = null;
 
-  public static getId() {
-    return RenderContainerBase.ids.push(true) - 1;
-  }
-  public static releaseId(rootId: number) {
-    RenderContainerBase.ids[rootId] = false;
-  }
-  public static compactIds() {
-    const idsCounter = RenderContainerBase.ids;
-    while (idsCounter.length && idsCounter[idsCounter.length - 1] === false) {
-      idsCounter.pop();
-    }
-  }
-
-  protected readonly rootId: number;
-  protected domContainer: Nullable<HTMLElement> = null;
+  private readonly rootId: number = rootId++;
 
   constructor(props: RenderContainerProps) {
     super(props);
-
-    RenderContainerBase.compactIds();
-    this.rootId = RenderContainerBase.getId();
 
     if (props.children) {
       this.mountContainer();
@@ -50,7 +32,12 @@ export class RenderContainerBase extends React.Component<RenderContainerProps> {
 
   public componentWillUnmount() {
     this.destroyContainer();
-    RenderContainerBase.releaseId(this.rootId);
+  }
+
+  public render() {
+    const Container = HAS_BUILTIN_PORTAL ? RenderContainerNative : RenderContainerFallback;
+
+    return <Container {...this.props} domContainer={this.domContainer} rootId={this.rootId} />;
   }
 
   private createContainer() {
@@ -96,73 +83,4 @@ export class RenderContainerBase extends React.Component<RenderContainerProps> {
   }
 }
 
-class RenderContainerNativePortal extends RenderContainerBase {
-  public static displayName = 'RenderContainer';
-
-  public render(): JSX.Element {
-    if (this.props.children) {
-      if (!this.domContainer) {
-        throw Error('There is no "this.domContainer"');
-      }
-
-      return (
-        <React.Fragment>
-          {this.props.anchor}
-          {ReactDOM.createPortal(this.props.children, this.domContainer)}
-          <NativePortal key="portal-ref" rt_rootID={this.rootId} />
-        </React.Fragment>
-      );
-    }
-
-    return <React.Fragment>{this.props.anchor}</React.Fragment>;
-  }
-}
-
-class RenderContainerCustomPortal extends RenderContainerBase {
-  public static displayName = 'RenderContainer';
-
-  public render(): JSX.Element {
-    return <CustomPortal rt_rootID={this.rootId}>{this.props.anchor}</CustomPortal>;
-  }
-
-  public componentDidMount() {
-    this.renderChild();
-  }
-
-  public componentDidUpdate() {
-    this.renderChild();
-  }
-
-  public componentWillUnmount() {
-    this.unmountChildren();
-    super.componentWillUnmount();
-  }
-
-  private renderChild() {
-    if (this.props.children) {
-      this.mountChildren();
-    } else {
-      this.unmountChildren();
-    }
-  }
-  private mountChildren() {
-    if (!this.domContainer) {
-      throw Error('There is no "this.domContainer"');
-    }
-    ReactDOM.unstable_renderSubtreeIntoContainer(
-      this,
-      <RootContainer rt_portalID={this.rootId}>{this.props.children}</RootContainer>,
-      this.domContainer,
-    );
-  }
-
-  private unmountChildren() {
-    if (this.domContainer) {
-      ReactDOM.unmountComponentAtNode(this.domContainer);
-    }
-  }
-}
-
-const HAS_BUILTIN_PORTAL = !!ReactDOM.createPortal;
-
-export default (HAS_BUILTIN_PORTAL ? RenderContainerNativePortal : RenderContainerCustomPortal);
+export default RenderContainer;
