+++ packages/retail-ui/components/RenderContainer/RenderContainerTypes.ts
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+export interface PortalPros {
+  rt_rootID: number;
+}
+
+export interface RenderContainerProps {
+  anchor?: React.ReactNode;
+  children?: React.ReactNode;
+}
